Move expander out of group module.
