+++ group/hash.go
@@ -1,10 +1,14 @@
 package group
 
-import "math/big"
+import (
+	"math/big"
+
+	"github.com/cloudflare/circl/expander"
+)
 
 // HashToField generates a set of elements {u1,..., uN} = Hash(b) where each
 // u in GF(p) and L is the security parameter.
-func HashToField(u []big.Int, b []byte, e Expander, p *big.Int, L uint) {
+func HashToField(u []big.Int, b []byte, e expander.Expander, p *big.Int, L uint) {
 	count := uint(len(u))
 	bytes := e.Expand(b, count*L)
 	for i := range u {
